+++ litemall-db/src/main/java/org/linlinjava/litemall/db/dao/GoodsProductMapper.java
@@ -3,6 +3,6 @@ package org.linlinjava.litemall.db.dao;
 import org.apache.ibatis.annotations.Param;
 
 public interface GoodsProductMapper {
-    int addStock(@Param("id") Integer id, @Param("num") Short num);
-    int reduceStock(@Param("id") Integer id, @Param("num") Short num);
+    int addStock(@Param("id") Integer id, @Param("num") Integer num);
+    int reduceStock(@Param("id") Integer id, @Param("num") Integer num);
 }
