+++ litemall-db/src/main/java/org/linlinjava/litemall/db/util/CouponUserConstant.java
@@ -1,8 +1,8 @@
 package org.linlinjava.litemall.db.util;
 
 public class CouponUserConstant {
-    public static final Short STATUS_USABLE = 0;
-    public static final Short STATUS_USED = 1;
-    public static final Short STATUS_EXPIRED = 2;
-    public static final Short STATUS_OUT = 3;
+    public static final Integer STATUS_USABLE = 0;
+    public static final Integer STATUS_USED = 1;
+    public static final Integer STATUS_EXPIRED = 2;
+    public static final Integer STATUS_OUT = 3;
 }
