+++ litemall-db/src/main/java/org/linlinjava/litemall/db/util/GrouponConstant.java
@@ -3,13 +3,13 @@ package org.linlinjava.litemall.db.util;
 import org.linlinjava.litemall.db.domain.LitemallOrder;
 
 public class GrouponConstant {
-    public static final Short RULE_STATUS_ON = 0;
-    public static final Short RULE_STATUS_DOWN_EXPIRE = 1;
-    public static final Short RULE_STATUS_DOWN_ADMIN = 2;
+    public static final Integer RULE_STATUS_ON = 0;
+    public static final Integer RULE_STATUS_DOWN_EXPIRE = 1;
+    public static final Integer RULE_STATUS_DOWN_ADMIN = 2;
 
 
-    public static final Short STATUS_NONE = 0;
-    public static final Short STATUS_ON = 1;
-    public static final Short STATUS_SUCCEED = 2;
-    public static final Short STATUS_FAIL = 3;
+    public static final Integer STATUS_NONE = 0;
+    public static final Integer STATUS_ON = 1;
+    public static final Integer STATUS_SUCCEED = 2;
+    public static final Integer STATUS_FAIL = 3;
 }
