+++ litemall-db/src/main/java/org/linlinjava/litemall/db/util/OrderUtil.java
@@ -21,16 +21,16 @@ import java.util.List;
  */
 public class OrderUtil {
 
-    public static final Short STATUS_CREATE = 101;
-    public static final Short STATUS_PAY = 201;
-    public static final Short STATUS_SHIP = 301;
-    public static final Short STATUS_CONFIRM = 401;
-    public static final Short STATUS_CANCEL = 102;
-    public static final Short STATUS_AUTO_CANCEL = 103;
-    public static final Short STATUS_ADMIN_CANCEL = 104;
-    public static final Short STATUS_REFUND = 202;
-    public static final Short STATUS_REFUND_CONFIRM = 203;
-    public static final Short STATUS_AUTO_CONFIRM = 402;
+    public static final Integer STATUS_CREATE = 101;
+    public static final Integer STATUS_PAY = 201;
+    public static final Integer STATUS_SHIP = 301;
+    public static final Integer STATUS_CONFIRM = 401;
+    public static final Integer STATUS_CANCEL = 102;
+    public static final Integer STATUS_AUTO_CANCEL = 103;
+    public static final Integer STATUS_ADMIN_CANCEL = 104;
+    public static final Integer STATUS_REFUND = 202;
+    public static final Integer STATUS_REFUND_CONFIRM = 203;
+    public static final Integer STATUS_AUTO_CONFIRM = 402;
 
     public static String orderStatusText(LitemallOrder order) {
         int status = order.getOrderStatus().intValue();
@@ -115,26 +115,26 @@ public class OrderUtil {
         return handleOption;
     }
 
-    public static List<Short> orderStatus(Integer showType) {
+    public static List<Integer> orderStatus(Integer showType) {
         // 全部订单
         if (showType == 0) {
             return null;
         }
 
-        List<Short> status = new ArrayList<Short>(2);
+        List<Integer> status = new ArrayList<Integer>(2);
 
         if (showType.equals(1)) {
             // 待付款订单
-            status.add((short) 101);
+            status.add(101);
         } else if (showType.equals(2)) {
             // 待发货订单
-            status.add((short) 201);
+            status.add(201);
         } else if (showType.equals(3)) {
             // 待收货订单
-            status.add((short) 301);
+            status.add(301);
         } else if (showType.equals(4)) {
             // 待评价订单
-            status.add((short) 401);
+            status.add(401);
 //            系统超时自动取消，此时应该不支持评价
 //            status.add((short)402);
         } else {
